+++ solo/src/commands/node.mjs
@@ -36,10 +36,12 @@ export class NodeCommand extends BaseCommand {
     if (!opts || !opts.downloader) throw new IllegalArgumentError('An instance of core/PackageDowner is required', opts.downloader)
     if (!opts || !opts.platformInstaller) throw new IllegalArgumentError('An instance of core/PlatformInstaller is required', opts.platformInstaller)
     if (!opts || !opts.keyManager) throw new IllegalArgumentError('An instance of core/KeyManager is required', opts.keyManager)
+    if (!opts || !opts.accountManager) throw new IllegalArgumentError('An instance of core/AccountManager is required', opts.accountManager)
 
     this.downloader = opts.downloader
     this.plaformInstaller = opts.platformInstaller
     this.keyManager = opts.keyManager
+    this.accountManager = opts.accountManager
   }
 
   async checkNetworkNodePod (namespace, nodeId) {
@@ -777,27 +779,28 @@ export class NodeCommand extends BaseCommand {
               })
             }
           })
-        .command({
-          command: 'jeromy',
-          desc: 'update all of the account keys',
-          builder: y => flags.setCommandFlags(y,
+          .command({
+            command: 'jeromy',
+            desc: 'update all of the account keys',
+            builder: y => flags.setCommandFlags(y,
               flags.namespace,
               flags.nodeIDs
-          ),
-          handler: argv => {
-            nodeCmd.logger.debug("==== Running 'node keys' ===")
-            nodeCmd.logger.debug(argv)
-            nodeCmd.jeromyTesting(argv)
-          }
-        })
+            ),
+            handler: argv => {
+              nodeCmd.logger.debug("==== Running 'node keys' ===")
+              nodeCmd.logger.debug(argv)
+              nodeCmd.jeromyTesting(argv)
+            }
+          })
           .demandCommand(1, 'Select a node command')
       }
     }
   }
-  jeromyTesting(argv) {
-    this.k8.jeromyTesting(argv).then( r => {
-          console.log(`result: ${JSON.stringify(r)}`)
-        }
+
+  jeromyTesting (argv) {
+    this.accountManager.jeromyTesting(argv).then(r => {
+      console.log(`result: ${JSON.stringify(r)}`)
+    }
     ).catch(err => {
       console.log(`error: ${JSON.stringify(err)}`)
     })
